fixed v32 wallet version and removed unused legacy price and options
